move a file
